test: add test update vector with chunks
