+++ tests/integration/crud/chunks/flow_kv.yml
@@ -1,0 +1,4 @@
+!Flow
+pods:
+  indexer:
+    uses: yaml/index_kv.yml
