+++ tests/integration/crud/chunks/flow_vector.yml
@@ -1,0 +1,4 @@
+!Flow
+pods:
+  indexer:
+    uses: yaml/index_vector.yml
