+++ tests/integration/crud/chunks/yaml/index_kv.yml
@@ -1,0 +1,31 @@
+!BinaryPbIndexer
+with:
+  index_filename: doc.gzip
+metas:
+  name: docIndexer
+  workspace: $JINA_TOPK_DIR
+
+requests:
+  on:
+    IndexRequest:
+      - !KVIndexDriver
+        with:
+          executor: docIndexer
+          traversal_paths: ['c']
+    SearchRequest:
+      - !KVSearchDriver
+        with:
+          executor: docIndexer
+          traversal_paths: ['r']
+    UpdateRequest:
+      - !KVIndexDriver
+        with:
+          method: update
+          executor: docIndexer
+          traversal_paths: [ 'c' ]
+    DeleteRequest:
+      - !KVIndexDriver
+        with:
+          method: delete
+          executor: docIndexer
+          traversal_paths: [ 'c' ]
