+++ tests/integration/crud/chunks/yaml/index_vector.yml
@@ -1,0 +1,34 @@
+!NumpyIndexer
+with:
+  index_filename: vec.gz
+  metric: cosine
+metas:
+  name: vecidx
+  workspace: $JINA_TOPK_DIR
+
+requests:
+  on:
+    UpdateRequest:
+      - !VectorIndexDriver
+        with:
+          method: update
+          executor: vecidx
+          traversal_paths: [ 'c' ]
+    DeleteRequest:
+      - !VectorIndexDriver
+        with:
+          method: delete
+          executor: vecidx
+          traversal_paths: [ 'c' ]
+    IndexRequest:
+      - !VectorIndexDriver
+        with:
+          executor: vecidx
+          traversal_paths: ['c']
+    SearchRequest:
+      - !VectorSearchDriver
+        with:
+          executor: vecidx
+          top_k: $JINA_TOPK
+          fill_embedding: True
+          traversal_paths: ['r']
